Fix imports; move connection logging to CameraIOInputsLogger
This better matches all other IO structures in code; there's no reason
the camera io shouldn't be the same.
